requirements.txt
a2wsgi==1.10.4
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.100.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
multidict==6.0.5
openai==1.10.0
pycparser==2.22
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.8.2
python-multipart==0.0.9
requests==2.31.0
retell-sdk==3.14.0
simple-websocket==1.0.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
twilio==8.12.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.21.1
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
